feat: add script for dev
