+++ package.json
@@ -9,7 +9,8 @@
     "clean": "gts clean",
     "compile": "tsc",
     "fix": "gts fix",
-    "start:dev": "nodemon src/index.ts"
+    "start:dev": "nodemon src/index.ts",
+    "start:prod": "npm run compile && node build/src/index.js"
   },
   "repository": {
     "type": "git",
@@ -23,7 +24,6 @@
   },
   "homepage": "https://github.com/anny-more/crud-api#readme",
   "devDependencies": {
-    "@types/express": "^4.17.17",
     "@types/node": "^14.11.2",
     "@types/uuid": "^9.0.2",
     "gts": "^3.1.1",
@@ -31,11 +31,8 @@
     "typescript": "^5.1.6"
   },
   "dependencies": {
-    "body-parser": "^1.20.2",
     "dotenv": "^16.3.1",
-    "express": "^4.18.2",
     "ts-node-dev": "^2.0.0",
-    "uuid": "^9.0.0",
-    "zod": "^3.21.4"
+    "uuid": "^9.0.0"
   }
 }
