+++ src/helpers/isInfoNotValid.ts
